requirements.txt
faker==19.6.0
pyyaml==6.0.1
click==8.1.7
openpyxl==3.1.2
pytest
fastparquet==2024.11.0
tqdm
pandas
pyarrow